Fixed compile warning messages on 64-bit.
git-svn-id: https://svn.code.sf.net/p/osp-toolkit/code/main@451 43700839-f8e2-0310-9689-8ed4146226e7

+++ include/osp/ospb64.h
@@ -27,9 +27,9 @@ extern "C"
 {
 #endif
 
-    int OSPPBase64Encode(unsigned const char *in,size_t inSize,
-                      unsigned char *out,size_t *outSize);
-    int OSPPBase64Decode(const char *in,size_t inSize,unsigned char *out,size_t *outSize);
+    int OSPPBase64Encode(unsigned const char *in,unsigned inSize,
+                      unsigned char *out,unsigned *outSize);
+    int OSPPBase64Decode(const char *in,unsigned inSize,unsigned char *out,unsigned *outSize);
 
 #ifdef __cplusplus
 }
